requirements.txt
langchain
langchain_community
faiss-cpu
langchain_chroma
langchain_huggingface
python-jobspy
langchain-google-genai
langgraph
bing-results-scraper
pdfplumber
lxml_html_clean
newspaper3k
chromadb
wikipedia
mistune
aiofiles
md2pdf
pypdf
gnews
langchain_openai
langchain_experimental
openai
MainContentExtractor
google-search-results
requests
bs4
streamlit_chat
duckduckgo_search